+++ src/test/kotlin/com/contentful/java/cma/model/Extensions.kt
@@ -1,0 +1,6 @@
+package com.contentful.java.cma.model
+
+fun CMASystem.setCreatedBy(link: CMALink): CMASystem {
+    this.createdBy = link
+    return this
+}